MR-76: activities routing started.
