+++ src/Components/Activities/Form/form.module.css
@@ -1,13 +1,3 @@
-.form-container {
-  position: absolute;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100vh;
-  background-color: #00000066;
-}
-
 .form {
   width: 700px;
   margin: 5% auto;
