--- src/Components/Activities/Modals/Delete/deleteModal.module.css
@@ -1,68 +1,0 @@
-.modal {
-  position: absolute;
-  z-index: 1;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #000000;
-  background-color: #00000066;
-}
-
-.modal-content {
-  display: flex;
-  flex-direction: column;
-  margin: 15% auto;
-  width: 313px;
-  min-height: 134px;
-  border: 1px solid #37386766;
-  border-radius: 4px;
-  background-color: #FFFFFF;
-  box-shadow: 0px 10px 23px #0000001f;
-  overflow: hidden;
-}
-
-.modal-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 16px;
-  color: #373867;
-  font-size: 1em;
-  font-weight: 700;
-}
-
-.modal-close {
-  cursor: pointer;
-}
-
-.modal-message {
-  padding: 0 16px;
-  font-weight: 400;
-  color: #373867;
-}
-
-.modal-buttons {
-  display: flex;
-  flex-direction: row-reverse;
-  padding: 16px;
-  gap: 6px;
-}
-
-.modal-buttons .btn {
-  width: 70px;
-  height: 36px;
-  border: 1px solid #373867;
-  border-radius: 5px;
-  color: #373867;
-  font-weight: 600;
-  cursor: pointer;
-}
-
-.btn-submit {
-  background-color: #AACE9B;
-}
-
-.btn-cancel {
-  background-color: #FFFFFF;
-}
