--- src/Components/Activities/Toast/toast.module.css
@@ -1,33 +1,0 @@
-.toast {
-  position: absolute;
-  display: flex;
-  align-items: center;
-  width: 313px;
-  height: 70px;
-  box-shadow: 0px 10px 23px rgba(0, 0, 0, 0.12);
-  border-radius: 5px;
-  color: #373867;
-  bottom: 35%;
-}
-
-.toast-content {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  padding: 0 20px;
-  font-size: 1em;
-  font-weight: 600;
-}
-
-.toast-closer {
-  cursor: pointer;
-}
-
-.toast-ok {
-  background-color: #AACE9B;
-}
-
-.toast-wrong {
-  background-color: #F13312;
-  color: #fff;
-}
